Tacocat passes the tests.
The Tacocat application passes all of the tests in app_tests.py. It is not a pretty webpage, but this was just a getting things running coding challenge and not a html one. It does seem to get an error at test_logged_in_menu if you fill out the database with a decent number of taco posts. It passes everything with a clean database though, so I am not sure what is going on with that.